Adding project_paths functionality.
This module contains the code to extract the template paths (and possible all other paths) needed for hte shunt system. Currently it will correctly extract the template paths from both folders that are children of the *root* shunt project file according to a given directory. The root folder is the closest ancestor htat has a Shuntfile. The Shuntfile may also contain explicitly added template paths which are returned first so as to be higher priority than implicit folder names